refactor(cs-bridge): merge errors.ts into types.ts
Move error classes (UserIdNotProvidedError, MessageParseError, AvatarFetchError)
into types.ts and delete errors.ts. Also removes parsePayload tests since
the function was deleted in Phase 1.2.

Phase 4.1 of kernel philosophy refactoring (9 → 8 files).
